Included functions to abstract the model interaction part.

+++ flaggy/doppio/api/controllers.py
@@ -8,6 +8,84 @@ from django.core.mail import send_mail, EmailMessage
 from doppio.api.emails import flaggy_email
 from doppio.api.proximity import coord_distance
 
+## MODELS RELATED METHODS ##
+
+def get_pk_user(pk):
+    try:
+        user = User.objects.get(pk=pk)
+        res = success()
+        res['user'] = user
+
+        return res
+    except Exception as inst:
+        return error("Error: %s " % inst)
+
+def get_fb_user(fb_id):
+    try:
+        user = User.objects.get(fb_id=fb_id)
+        res = success()
+        res['user'] = user
+
+        return res
+    except Exception as inst:
+        return error("Error: %s " %inst)
+
+def get_follow_request(f_er, f_ing):
+    try:
+        req = FollowPending.objects.get(follower_p_id=f_er, following_p_id=f_ing)
+        res = success()
+        res['req'] = req
+
+        return res
+
+    except Exception as inst:
+        return error("Error: %s " % inst)
+
+def get_follow(f_er, f_ing):
+    try:
+        follow = Follow.objects.get(follower_id=follower, following_id=followed)
+        res = success()
+        res['follow'] = follow
+
+        return res
+    except Exception as inst:
+        return error("Error: %s " % inst)
+
+
+def follow_hash(u_id1, u_id2):
+    return hashlib.sha224("%s&%s" % (u_id1, u_id2)).hexdigest()
+
+## SUCCESS/ERROR RESPONSES ##
+
+def success(msg):
+    return {'status': 'success', 'msg': msg}
+
+
+def error(msg):
+    return {'status': 'error', 'msg': msg}
+
+
+## HELPERS ##
+## Here will be the functions that are not directly mapped to a view ##
+
+def empty_str(s):
+    return s == "" or s is None
+
+def last_check_in(user_id):
+    try:
+        checkin = CheckIn.objects.filter(u_id=user_id).latest('when')
+        return {
+            'lng': str(checkin.longitude),
+            'lat': str(checkin.latitude),
+            'when': str(checkin.when),
+            'comment': str(checkin.comment)
+        }
+    except CheckIn.DoesNotExist:
+        return None
+
+
+## REQUEST HANDLERS ##
+
 def __add_user(f_n, l_n, fb, twitter, email):
     try:
         u = User(
@@ -44,8 +122,13 @@ def __add_user(f_n, l_n, fb, twitter, email):
 ## FOLLOW AND UNFOLLOW ##
 
 def __add_follow(follower, followed_fb):
-    try:
-        f_er = User.objects.get(pk=follower)
+    follower = get_pk_user(follower)
+    followed = get_fb_user(followed_fb)
+
+    if(user['status'] == "error"): res = follower
+
+    elif(user['status'] == "success"):
+        f_er = follower['user']
 
         if(int(f_er.fb_id) == int(followed_fb)):
             return error("Why do you want to follow yourself?")
@@ -53,10 +136,14 @@ def __add_follow(follower, followed_fb):
         email_info = { }
         email_info["follower"] = f_er.fname
 
-        if verify_user(followed_fb):
+        if(followed['status'] == "error"):
+            res = error("Facebook user %s not in our database" % followed_fb)
+
+        elif (followed['status'] == "success"):
             # Send email to existing user. Add to follow pending table
-            f_ed = User.objects.get(fb_id=followed_fb)
-            k = hashlib.sha224("%s&%s" % (f_er.pk, f_ed.pk)).hexdigest()
+            f_ed = followed['user']
+
+            k = follow_hash(f_er.pk, f_ed.pk)
             
             email_info["key"] = k
             email_info["template"] = "follow"
@@ -67,34 +154,34 @@ def __add_follow(follower, followed_fb):
             if (mail_status == "success"):
                 f = FollowPending(follower_p=f_er, following_p=f_ed, secure_key=k)
                 f.save()
+                
                 res = success("Request sent to %s." % f_ed.fname)
+
             elif(mail_status == "error"):
                 res = error("Failed to send email request.")
 
-        else:
-            res = error("Facebook user %s not in our database" % followed_fb)
-
-        return res
-    
-    except User.DoesNotExist:
-        return error("Follower Does not exist.")
+    return res
 
 
 def __unfollow(follower, followed):
-    try:
-        f = FollowPending.objects.get(follower_p_id=follower, following_p_id=followed)
-        if f.approve:
+    req = get_follow_request(follower, followed)
+    follow = get_follow(follower, followed)
+
+    if(req['status'] == "error"): res = req
+
+    elif(req['status'] == "success"):
+
+        f = req['req']
+
+        if (f.approve and follow['status'] == "success"):
             f.approve = False
-            follow = Follow.objects.get(follower_id=follower, following_id=followed)
+            curr_follow = follow['follow']
             f.delete()
-            follow.delete()
-            return success("Successfully unfollowed.")
+            curr_follow.delete()
+            res =success("Successfully unfollowed.")
         else:
-            return error("Already unfollowed.")
-    except FollowPending.DoesNotExist:
-        return error("No such connection exists.")
-    except:
-        return error("Error. Could not unfollow.")
+            res = error("Already unfollowed.")
+    return res
 
 
 def __approve_request(k):
@@ -172,8 +259,11 @@ def __approved_request():
     return res
 
 def __retrieve_f_request(follower_id, following_id):
-    try:
-        f = FollowPending.objects.get(follower_p_id=follower_id, following_p_id=following_id)
+    f = get_follow_request(follower_id, following_id)
+
+    if(f['status'] == "error"): res = f
+
+    elif(f['status'] == "success"):
         res = success("Found request.")
         f_dict = { }
         f_dict["p_id"] = int(f.p_id)
@@ -184,10 +274,7 @@ def __retrieve_f_request(follower_id, following_id):
 
         res["request"] = f_dict
 
-        return res
-
-    except FollowPending.DoesNotExist:
-        return error("Such a follow request does not exist.")
+    return error("Such a follow request does not exist.")
 
 def __unapproved_requests():
     res = success("Found all unapproved requests.")
@@ -262,15 +349,8 @@ def __show_checkins(u_id):
 
     return res
 
-def get_user(u_id):
-    try:
-        user = User.objects.get(pk=u_id)
-        return user
-    except User.DoesNotExist:
-        return None
-
 def __nearby(u_id):
-    if(get_user(u_id) is None):
+    if(get_pk_user(u_id) is None):
         return error("User %s does not exist" % u_id)
     else:
         user = User.objects.get(pk=u_id)
@@ -325,47 +405,6 @@ def __nearby(u_id):
 
 #    except Exception as inst:
 #        error("Error retrieving followers. %s" % inst)
-
-
-
-## RESPONSES ##
-
-def success(msg):
-    return {'status': 'success', 'msg': msg}
-
-
-def error(msg):
-    return {'status': 'error', 'msg': msg}
-
-
-## HELPERS ##
-## Here will be the functions that are not directly mapped to a view ##
-
-def empty_str(s):
-    return s == "" or s is None
-
-
-def verify_user(value):
-    try:
-        User.objects.get(fb_id=value)
-        return True
-    except User.MultipleObjectsReturned:
-        return True
-    except User.DoesNotExist:
-        return False
-
-
-def last_check_in(user_id):
-    try:
-        checkin = CheckIn.objects.filter(u_id=user_id).latest('when')
-        return {
-            'lng': str(checkin.longitude),
-            'lat': str(checkin.latitude),
-            'when': str(checkin.when),
-            'comment': str(checkin.comment)
-        }
-    except CheckIn.DoesNotExist:
-        return None
